meta_data_insertion in nordic_to_nlbpub
insert_metadata - added a method that checks if the metadata has been inserted from before - that is in nordic_to_nlbpub and if so only update the dc:identifier for the specific format else insert as normal. In nordic_to_nlbpub added a method that inserts the data by calling existing methods in utils and another one that reorder metadata attributes caused by beatifulsoup.